+++ clang/include/clang/Format/Format.h
@@ -3001,6 +3001,29 @@ struct FormatStyle {
   /// \version 11
   bool ObjCBreakBeforeNestedBlockParam;
 
+  /// The order in which property attributes should appear, such as:
+  ///
+  ///   * class
+  ///   * direct
+  ///   * atomic, nonatomic
+  ///   * assign, retain, strong, copy, weak, unsafe_unretained
+  ///   * readonly, readwrite
+  ///   * getter, setter
+  ///   * nullable, nonnull, null_resettable, null_unspecified
+  ///
+  /// Attributes in code will be sorted in the order specified. Any attributes
+  /// encountered that are not mentioned in this array will be sorted last, in
+  /// stable order. Duplicate attributes will be removed, but no other conflict
+  /// checking is performed. A leading or trailing comment is allowed to the
+  /// whole set, but comments encountered between attributes will leave the
+  /// entire set untouched.
+  ///
+  /// \code{.yaml}
+  ///   ObjCPropertyAttributeOrder: [nonatomic, strong, readwrite, nullable]
+  /// \endcode
+  /// \version 17
+  std::vector<std::string> ObjCPropertyAttributeOrder;
+
   /// Add a space after ``@property`` in Objective-C, i.e. use
   /// ``@property (readonly)`` instead of ``@property(readonly)``.
   /// \version 3.7
@@ -4432,6 +4455,7 @@ struct FormatStyle {
            ObjCBlockIndentWidth == R.ObjCBlockIndentWidth &&
            ObjCBreakBeforeNestedBlockParam ==
                R.ObjCBreakBeforeNestedBlockParam &&
+           ObjCPropertyAttributeOrder == R.ObjCPropertyAttributeOrder &&
            ObjCSpaceAfterProperty == R.ObjCSpaceAfterProperty &&
            ObjCSpaceBeforeProtocolList == R.ObjCSpaceBeforeProtocolList &&
            PackConstructorInitializers == R.PackConstructorInitializers &&